Report the capture actor first in whoami, then hosted auth
whoami was hosted-only, so init --principal-* / user_config showed up in
status but not in whoami. Print the local capture actor first and keep
hosted authentication as a second stanza, including when unauthenticated.

Fixes #1437
